requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==2.1.3
Flask-JWT==0.3.2
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
psycopg2==2.9.6
PyJWT==1.4.2
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.27
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==2.2.2
zipp==3.17.0
